Change the calls and incomings structure
Rename the call id to call, and the incoming id to inc, so it is unique
within the database and thus can be used in natural join and join ..
using.

Remove covered and open variables and replace with a status. Different
statuses can be added to call_statuses table, but at the moment only
Open, Covered and Ignored is supported.